requirements.txt
pandas
scikit-learn
python-dotenv
perplexityai
flask
joblib